backend/python/requirements.txt
torch
torchvision
timm
fastapi
uvicorn
pillow
numpy
